Gradle Buildscript.
CoFHLib is packed in built jars of Core.
AccessTransformer has been nuked. Add things back as they are needed. If
there is a getter / setter, use it OR submit a PR for a getter / setter
if it is usefull.
Avid using * in access transformers everywhere. Try and make them as
less invaisve as possible.

+++ src/main/java/cofh/core/entity/EntityCoFHFishHook.java
@@ -54,7 +54,7 @@ public class EntityCoFHFishHook extends EntityFishHook {
         onEntityUpdate();
 
         if (this.worldObj.isRemote) {
-            int i = ((Integer) this.getDataManager().get(DATA_HOOKED_ENTITY)).intValue();
+            int i = this.getDataManager().get(DATA_HOOKED_ENTITY);
 
             if (i > 0 && this.caughtEntity == null) {
                 this.caughtEntity = this.worldObj.getEntityByID(i - 1);
